lambda/requirements.txt
# AWS SDK for Python
boto3>=1.35.0,<2.0.0

# Image processing
Pillow>=11.0.0,<12.0.0

# HTTP client for backend callbacks
urllib3>=2.0.0,<3.0.0

# Environment variable management
python-dotenv>=1.0.0,<2.0.0

# Type hints support
typing-extensions>=4.12.0
